FREEPBX-13470 No default lot.
Issue, on a clean (new) install there is no lots. This causes whoops errors on apply_config.

Change: parking_get initialize the results as an array, in the case of no lots an empty array is returned.
Fix: Add the default lot logic back in to the install file to create/prune default lots.

QA TESTING:
Install parking drop any existing tables (parkplus). On a clean install apply_config and it will fail.
After Update a default lot will be present preventing any errors.
The array fix is suplimental There should never NOT be any lots on a proper install.